moves card window handling to own module
